+++ main.py
@@ -1,9 +1,3 @@
-"""预测中国的假期安排的模块。
-前置模块：zhdate
-用法：CalculationUtil.假期名(年份)
-输出：<假期开始日期>, <假期结束日期>, [调休日期1(若有)], [调休日期2(若有)]
-"""
-
 """
 Holiday Predictor / 假期预测器 - 基于 Python 的调休预测工具。
 
@@ -21,264 +15,19 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
-__all__ = ["CalculationUtil", "calculation"]
+__all__ = ["calculation"]
 
 import sys
 
-from typing import List, Optional, Tuple
+from typing import Optional, Tuple, List, ClassVar
 
 import datetime
 
 from zhdate import ZhDate  # type: ignore
 
-from data import Data
+from holiday_predictor import CalculationUtil
 
-
-class CalculationUtil:
-    """计算假期安排所需的类。
-    用法：CalculationUtil.假期名(年份)
-    输出：<假期开始日期>, <假期结束日期>, [调休日期1(若有)], [调休日期2(若有)]
-    """
-
-    @staticmethod
-    def national_day(year: int) -> Optional[Tuple[
-        datetime.datetime, datetime.datetime, Optional[datetime.datetime],
-        Optional[datetime.datetime]
-    ]]:
-        if year < 1949:
-            return None
-        # 下述代码从传入参数中获取必要信息。
-
-        # 这两个变量定义调休的日期。在随后会被修改，用 "None" 来判定是否有调休出现。
-        lieu_2: Optional[datetime.datetime] = None
-        # 下述代码对中秋节、国庆节的日期进行运算。
-        mid_date: datetime.datetime = ZhDate(
-            year, 8, 15
-        ).to_datetime()  # 该变量是中秋节的日期。
-        nati_date = datetime.datetime(
-            year, 10, 1
-        )  # 该变量是国庆节开始的日期。
-        mid_dateofweek: int = mid_date.weekday()  # 该变量表明中秋节在周几。
-        # national_day_date_of_week = national_day_date.weekday()
-        is_mid_in_nati: bool = \
-            False  # 该变量用于确定中秋节是否与国庆节相连或被包含其中。
-        hld_startdate: datetime.datetime = nati_date
-
-        # 下述代码对假期开始的日期进行运算。
-        if (
-                mid_dateofweek >= 4
-                and datetime.timedelta(days=-3)
-                <= mid_date - nati_date
-                <= datetime.timedelta(days=0)
-        ) or (mid_date == nati_date):
-            hld_startdate = mid_date
-            is_mid_in_nati = True
-        if datetime.timedelta(days=0) \
-                < mid_date - nati_date \
-                <= datetime.timedelta(days=7):
-            hld_startdate = nati_date
-            is_mid_in_nati = True
-        if not is_mid_in_nati:
-            hld_startdate = nati_date
-
-        # 下述代码对假期结束的日期进行运算。
-        hld_days: int = 7 + is_mid_in_nati  # 该变量指示了假期的长度。
-        hld_start_dateofweek: int = hld_startdate.weekday()  # 该变量表明假期在周几开始。
-        hld_enddate = hld_startdate + datetime.timedelta(
-            days=hld_days - 1, hours=23, minutes=59, seconds=59
-        )  # 该变量是假期结束的日期。
-        hld_end_dateofweek = hld_enddate.weekday()  # 该变量表明假期在周几结束。
-
-        # 下述代码对调休的日期进行运算。
-        if hld_end_dateofweek == 4:
-            lieu_1 = hld_enddate + datetime.timedelta(seconds=1)
-            lieu_2 = hld_enddate + datetime.timedelta(days=2)
-        elif hld_start_dateofweek == 0:
-            lieu_1 = hld_startdate - datetime.timedelta(days=2)
-            lieu_2 = hld_startdate - datetime.timedelta(seconds=1)
-        elif hld_start_dateofweek == 6 and hld_days == 8:
-            lieu_1 = hld_startdate - datetime.timedelta(days=1)
-        else:
-            lieu_1 = hld_startdate - datetime.timedelta(
-                days=hld_start_dateofweek + 1
-            )
-            lieu_2 = hld_enddate + datetime.timedelta(days=5 - hld_end_dateofweek)
-        return hld_startdate, hld_enddate, lieu_1, lieu_2
-
-    @staticmethod
-    def new_year(year: int) -> Tuple[
-        datetime.datetime, datetime.datetime, Optional[datetime.datetime]
-    ]:
-        new_year_date: datetime.datetime = datetime.datetime(
-            year=year, month=1, day=1
-        )
-        new_year_dateofweek: int = new_year_date.weekday()
-        hld_startdate: datetime.datetime = \
-            new_year_date + datetime.timedelta(
-                Data.HLD_3DAYS_START_DELTA_DAY[new_year_dateofweek]
-            )
-
-        hld_enddate: datetime.datetime = \
-            new_year_date + datetime.timedelta(
-                Data.HLD_3DAYS_END_DELTA_DAY[new_year_dateofweek]
-            )
-
-        lieu_1: Optional[datetime.datetime] = \
-            new_year_date + datetime.timedelta(
-                Data.HLD_3DAYS_LIEU1_DELTA_DAY[new_year_dateofweek]  # type: ignore
-            ) if Data.HLD_3DAYS_LIEU1_DELTA_DAY[new_year_dateofweek] is not None \
-                else None  # type: ignore
-        return hld_startdate, hld_enddate, lieu_1
-
-    @staticmethod
-    def spring_festival(year: int) -> Tuple[
-        datetime.datetime, datetime.datetime, datetime.datetime,
-        datetime.datetime
-    ]:
-        spring_festival_date: datetime.datetime = \
-            ZhDate(year, 1, 1).to_datetime()
-        spring_festival_dateofweek: int = spring_festival_date.weekday()
-        # 下述代码对春节假期的调休进行运算。
-        hld_startdate: datetime.datetime = \
-            spring_festival_date + datetime.timedelta(
-                Data.SPRING_FESTIVAL_START_DELTA_DAY[spring_festival_dateofweek]
-            )
-        hld_enddate: datetime.datetime = \
-            spring_festival_date + datetime.timedelta(
-                Data.SPRING_FESTIVAL_END_DELTA_DAY[spring_festival_dateofweek]
-            )
-        lieu_1: datetime.datetime = \
-            spring_festival_date + datetime.timedelta(
-                Data.SPRING_FESTIVAL_LIEU1_DELTA_DAY[spring_festival_dateofweek]
-            )
-        lieu_2: datetime.datetime = \
-            spring_festival_date + datetime.timedelta(
-                Data.SPRING_FESTIVAL_LIEU2_DELTA_DAY[spring_festival_dateofweek]
-            )
-        return hld_startdate, hld_enddate, lieu_1, lieu_2
-
-    @staticmethod
-    def qing_ming(year: int) -> Tuple[
-        datetime.datetime, datetime.datetime, Optional[datetime.datetime]
-    ]:
-        qing_ming_date: datetime.datetime = datetime.datetime(
-            year=year, month=4,
-            day=5 if year % 4 or (year % 400 and not year % 100) else 4
-        )
-        qing_ming_dateofweek: int = qing_ming_date.weekday()
-        # 下述代码对清明假期的调休进行运算。
-        hld_startdate: datetime.datetime = \
-            qing_ming_date + datetime.timedelta(
-                Data.HLD_3DAYS_START_DELTA_DAY[qing_ming_dateofweek]
-            )
-
-        hld_enddate: datetime.datetime = \
-            qing_ming_date + datetime.timedelta(
-                Data.HLD_3DAYS_END_DELTA_DAY[qing_ming_dateofweek]
-            )
-
-        lieu_1: Optional[datetime.datetime] = \
-            qing_ming_date + datetime.timedelta(
-                Data.HLD_3DAYS_LIEU1_DELTA_DAY[qing_ming_dateofweek]  # type: ignore
-            ) if Data.HLD_3DAYS_LIEU1_DELTA_DAY[qing_ming_dateofweek] is not None \
-                else None  # type: ignore
-        return hld_startdate, hld_enddate, lieu_1
-
-    @staticmethod
-    def duan_wu(year: int) -> Tuple[
-        datetime.datetime, datetime.datetime, Optional[datetime.datetime]
-    ]:
-        duan_wu_date: datetime.datetime = ZhDate(year, 5, 5).to_datetime()
-        duan_wu_dateofweek: int = duan_wu_date.weekday()
-        # 下述代码对端午假期的调休进行运算。
-        hld_startdate: datetime.datetime = \
-            duan_wu_date + datetime.timedelta(
-                Data.HLD_3DAYS_START_DELTA_DAY[duan_wu_dateofweek]
-            )
-
-        hld_enddate: datetime.datetime = \
-            duan_wu_date + datetime.timedelta(
-                Data.HLD_3DAYS_END_DELTA_DAY[duan_wu_dateofweek]
-            )
-
-        lieu_1: Optional[datetime.datetime] = \
-            duan_wu_date + datetime.timedelta(
-                Data.HLD_3DAYS_LIEU1_DELTA_DAY[duan_wu_dateofweek]  # type: ignore
-            ) if Data.HLD_3DAYS_LIEU1_DELTA_DAY[duan_wu_dateofweek] is not None \
-                else None  # type: ignore
-        return hld_startdate, hld_enddate, lieu_1
-
-    @staticmethod
-    def international_labours_day(year: int) -> Optional[Tuple[
-        datetime.datetime, datetime.datetime, datetime.datetime,
-        datetime.datetime
-    ]]:
-        if year < 1890:
-            return None
-        international_labours_day_date: datetime.datetime = \
-            datetime.datetime(year=year, month=5, day=1)
-        international_labours_day_dateofweek: int = \
-            international_labours_day_date.weekday()
-        # 下述代码对五一假期的调休进行运算。
-        hld_startdate: datetime.datetime = \
-            international_labours_day_date + datetime.timedelta(
-                Data.INTERNATIONAL_LABOURS_DAY_START_DELTA_DAY[
-                    international_labours_day_dateofweek
-                ]
-            )
-        hld_enddate: datetime.datetime = \
-            international_labours_day_date + datetime.timedelta(
-                Data.INTERNATIONAL_LABOURS_DAY_END_DELTA_DAY[
-                    international_labours_day_dateofweek
-                ]
-            )
-        lieu_1: datetime.datetime = \
-            international_labours_day_date + datetime.timedelta(
-                Data.INTERNATIONAL_LABOURS_DAY_LIEU1_DELTA_DAY[
-                    international_labours_day_dateofweek
-                ]
-            )
-        lieu_2: datetime.datetime = \
-            international_labours_day_date + datetime.timedelta(
-                Data.INTERNATIONAL_LABOURS_DAY_LIEU2_DELTA_DAY[
-                    international_labours_day_dateofweek
-                ]
-            )
-        return hld_startdate, hld_enddate, lieu_1, lieu_2
-
-    @classmethod
-    def mid_autumn(cls, year: int) -> Tuple[
-        datetime.datetime, datetime.datetime, Optional[datetime.datetime],
-        Optional[datetime.datetime]
-    ]:
-        mid_autumn_date: datetime.datetime = \
-            ZhDate(year, 8, 15).to_datetime()
-        mid_autumn_dateofweek: int = mid_autumn_date.weekday()
-        # 下述代码对中秋假期的调休进行运算。
-        if year >= 1949:
-            national_day_result: Optional[Tuple[
-                datetime.datetime, datetime.datetime,
-                Optional[datetime.datetime], Optional[datetime.datetime]
-            ]] = cls.national_day(year)
-            if national_day_result is not None and \
-                    (national_day_result[1] - national_day_result[0]).days > 6:
-                return national_day_result
-        hld_startdate: datetime.datetime = \
-            mid_autumn_date + datetime.timedelta(
-                Data.HLD_3DAYS_START_DELTA_DAY[mid_autumn_dateofweek]
-            )
-
-        hld_enddate: datetime.datetime = \
-            mid_autumn_date + datetime.timedelta(
-                Data.HLD_3DAYS_END_DELTA_DAY[mid_autumn_dateofweek]
-            )
-
-        lieu_1: Optional[datetime.datetime] = \
-            mid_autumn_date + datetime.timedelta(
-                Data.HLD_3DAYS_LIEU1_DELTA_DAY[mid_autumn_dateofweek]  # type: ignore
-            ) if Data.HLD_3DAYS_LIEU1_DELTA_DAY[mid_autumn_dateofweek] is not None \
-                else None  # type: ignore
-        return hld_startdate, hld_enddate, lieu_1, None
+from text_info import TextInfo
 
 
 def calculation(given_list: List[str]) -> str:
@@ -314,7 +63,7 @@ def calculation(given_list: List[str]) -> str:
             CalculationUtil.new_year(forecast_year)
     elif holiday_name in ("--spring-festival", "-sf"):  # 该部分用于处理春节假期的调休预测。
         if '--do-not-output-notes' not in given_list:
-            print(Data.spring_festival_note)
+            print(TextInfo.spring_festival_note)
         spring_festival_result: Tuple[
             datetime.datetime, datetime.datetime, datetime.datetime,
             datetime.datetime
@@ -328,7 +77,7 @@ def calculation(given_list: List[str]) -> str:
             CalculationUtil.duan_wu(forecast_year)
     elif holiday_name in ("--international-labours-day", "-ild"):  # 该部分用于处理五一假期的调休预测。
         if '--do-not-output-notes' not in given_list:
-            print(Data.international_labours_day_note)
+            print(TextInfo.international_labours_day_note)
         international_labours_day_result: Optional[Tuple[
             datetime.datetime, datetime.datetime, datetime.datetime,
             datetime.datetime
@@ -350,31 +99,31 @@ def calculation(given_list: List[str]) -> str:
         return ("假期由 {start}（星期{start_dateofweek}）起，直到 {end}（星期{end_dateofweek}），共 {day} 天。"
                 "调休时间为 {lieu1}（星期{lieu1_dateofweek}）和 {lieu2}（星期{lieu2_dateofweek}）。".format(
             start=hld_startdate,
-            start_dateofweek=Data.INT_TO_WEEKDAY[hld_startdate.weekday()],
+            start_dateofweek=TextInfo.INT_TO_WEEKDAY[hld_startdate.weekday()],
             end=hld_enddate,
-            end_dateofweek=Data.INT_TO_WEEKDAY[hld_enddate.weekday()],
+            end_dateofweek=TextInfo.INT_TO_WEEKDAY[hld_enddate.weekday()],
             lieu1=lieu_1.date(),
-            lieu1_dateofweek=Data.INT_TO_WEEKDAY[lieu_1.weekday()],
+            lieu1_dateofweek=TextInfo.INT_TO_WEEKDAY[lieu_1.weekday()],
             lieu2=lieu_2.date(),
-            lieu2_dateofweek=Data.INT_TO_WEEKDAY[lieu_2.weekday()],
+            lieu2_dateofweek=TextInfo.INT_TO_WEEKDAY[lieu_2.weekday()],
             day=hld_days
         ))
     if lieu_1 is not None:
         return ("假期由 {start}（星期{start_dateofweek}）起，直到 {end}（星期{end_dateofweek}），共 {day} 天。"
                 "调休时间为 {lieu}（星期{lieu_dateofweek}）。".format(
             start=hld_startdate,
-            start_dateofweek=Data.INT_TO_WEEKDAY[hld_startdate.weekday()],
+            start_dateofweek=TextInfo.INT_TO_WEEKDAY[hld_startdate.weekday()],
             end=hld_enddate,
-            end_dateofweek=Data.INT_TO_WEEKDAY[hld_enddate.weekday()],
+            end_dateofweek=TextInfo.INT_TO_WEEKDAY[hld_enddate.weekday()],
             lieu=lieu_1.date(),
-            lieu_dateofweek=Data.INT_TO_WEEKDAY[lieu_1.weekday()],
+            lieu_dateofweek=TextInfo.INT_TO_WEEKDAY[lieu_1.weekday()],
             day=hld_days
         ))
     return "假期由 {start}（星期{start_dateofweek}）起，直到 {end}（星期{end_dateofweek}），共 {day} 天。".format(
         start=hld_startdate,
-        start_dateofweek=Data.INT_TO_WEEKDAY[hld_startdate.weekday()],
+        start_dateofweek=TextInfo.INT_TO_WEEKDAY[hld_startdate.weekday()],
         end=hld_enddate,
-        end_dateofweek=Data.INT_TO_WEEKDAY[hld_enddate.weekday()],
+        end_dateofweek=TextInfo.INT_TO_WEEKDAY[hld_enddate.weekday()],
         day=hld_days
     )
 
@@ -417,7 +166,7 @@ if __name__ == "__main__" and len(sys.argv) < 2:
             execution_flag = True  # pylint: disable = C0103
 
         if input_list[0] == "help":
-            print(Data.help)
+            print(TextInfo.help)
             execution_flag = True  # pylint: disable = C0103
 
         if input_list[0] in ("exit", "quit"):
@@ -454,7 +203,7 @@ elif __name__ == "__main__" and len(sys.argv) > 1:
             sys.exit(2)
 
     elif input_list[0] == "help":
-        print(Data.help)
+        print(TextInfo.help)
         sys.exit(0)
 
     else:
